add extensible services
This commit adds extensible services to Embeddable. Additional services
can be defined by subclassing `Embeddable::Service` and providing an
`id` method. See `default_services.rb` for examples.

In addition to providing extensibility, the service lookup should be
O(1) instead of O(N), which does not matter for the small number of
services available now, but could matter if it is ever extended to
include hundreds of services.

The domain handling is also much more robust as the URIs are parsed
instead of using a regexp. The services can now handle all protocols,
all subdomains, and querystring parameters in any order.

Backwards compatibility is preserved with the exception of defining a
method with the same name as the embeddable which returns its service.